streamlit/requirements.txt
streamlit==1.20.0
tensorflow==2.12.0
pandas==2.0.1
streamlit-option-menu
streamlit-drawable-canvas
Pillow
numpy
scikit-learn
imbalanced-learn
ucimlrepo